Support autodetecting partition in `forXXX` (#133)
This expands a bit on the partition support added in #129. It helps CDK
users who need to do `forAccount("123456789012")` so that the partition
in the ARN is correct; however, it doesn't provide a more general
solution since `defaultPartition` is still hardcoded to `"aws"` for the
standard version of the library.

Because the `forXXX` methods accept a variable-length argument in the
last position, adding a `partition?: string` parameter would require
placing it before that argument, becoming an API-breaking change for a
rarer use-case.

A workaround is to use the `for()` in non-standard partitions and to
supply the ARN directly. But this will make it at least a little easier
for `cdk-iam-floyd` users (though, they have `forCdkPrincipal` so maybe
it doesn't help a ton).